+++ docs/examples/ex_140_time_delta.py
@@ -1,4 +1,4 @@
-from pytermor.render import RendererManager, SGRRenderer
+from pytermor.render import RendererManager, SgrRenderer
 from pytermor.util import time_delta
 
 seconds_list = [2, 10, 60, 2700, 32340, 273600, 4752000, 864000000]
@@ -7,7 +7,7 @@ max_len_list = [3, 6, 10]
 for max_len in max_len_list:
     formatter = time_delta.registry.find_matching(max_len)
 
-RendererManager.set_up(SGRRenderer)
+RendererManager.set_up(SgrRenderer)
 for seconds in seconds_list:
     for max_len in max_len_list:
         formatter = time_delta.registry.get_by_max_len(max_len)